% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teachboostr.R
\docType{package}
\name{deanslistr}
\alias{deanslistr}
\alias{deanslistr-package}
\title{teachboostr}
\description{
teachboostr is an R API for the TeachBoost's RESTful API
}
\section{Main functions}{

\code{teachboost_api()}
}

\section{Package context}{

For getting data out of TeachBoost for analysis as well as
transformation and loading into your various data stores.
}
\keyword{internal}

