% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_content}
\alias{extract_content}
\title{Extracts content of a a collection of teachboost responses returned by \link{get_tb()}}
\usage{
extract_content(x)
}
\arguments{
\item{x}{the output (a list of response objects)}
}
\value{
a list of the content portions from a collection of TeachBoost response objects
}
\description{
Extracts content of a a collection of teachboost responses returned by \link{get_tb()}
}

