% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tb.R
\name{get_tb}
\alias{get_tb}
\title{Get and collect responses from the TeachBoost API.}
\usage{
get_tb(endpoint, key, ...)
}
\arguments{
\item{endpoint}{the TeachBoost API endpoint to sent a GET request to.}

\item{key}{your organizations unique TeachBoost API key.  If one is not supplied the function will
look in the R enivronment for a \code{TB_KEY} key and retriev it's value. You can place your key in
an \code{.Renviron} file.}

\item{...}{parameters passed on to \code{\link{teachboost_api}}}
}
\value{
a list of teachboost response objects
}
\description{
A wrapper around \code{\link{teachboost_api}}, which will return as many response
objects as needed (given any constraints passed on to \code{\link{teachboost_api}})
}
\examples{
\dontrun{
x <- get_teachboost("users")

x <- get_teachboost("forms", limit = 1200)
}
}

