% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_to_df}
\alias{list_to_df}
\title{Translate (possibly nested) Teachboost List into a dataframe}
\usage{
list_to_df(.data)
}
\arguments{
\item{.data}{parsed list returned from \code{\link{parse_content}}}
}
\value{
a single row data frame
}
\description{
This simply takes a list constructed from \code{\link{parse_content}} and create's a
data frame (of the \code{tibble} variety).  It will attempt to to keep simplify the data frame as much
as possible, but not too much.  It tries to keep nested lists in a nested format.
}

