% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_content}
\alias{parse_content}
\title{Parse the JSON content of a TeachBoost response into a a list}
\usage{
parse_content(content)
}
\arguments{
\item{content}{JSON string (from Teachboost response's content) to be parsed.}
}
\value{
a (possibly nested) list
}
\description{
Note that this function does very minimal simplificaton.
For example, nested JSON is returned as nested lists (rather than a flattened list).
}

