% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teachboost_api.R
\name{teachboost_api}
\alias{teachboost_api}
\title{R API for TeachBoost's RESTful API}
\usage{
teachboost_api(endpoint, key, parsed = FALSE, ...)
}
\arguments{
\item{endpoint}{the endpoint from which data is requested}

\item{key}{your organizaitons TeachBoost API token}

\item{parsed}{wheterh the whether the response content should be parsed to a (possibly flattened) data.frame (`TRUE`) or
the whether the raw JSON us returned (`FALSE`).}

\item{...}{Not used}
}
\value{
a teadchboostr object, which is a list containing the response, endpoint path,
and returned content
}
\description{
R API for TeachBoost's RESTful API
}
\details{
This is a workhourse function that wraps the HTTP RESTful API that TeachBoos has implemented.
You will need your API key, which you'll need to request from TeachBoost.
}

