% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_cols.R
\name{unnest_list_cols}
\alias{unnest_list_cols}
\title{Find and unnest any list-columns in a data frame.}
\usage{
unnest_list_cols(.data, id_col = "id")
}
\arguments{
\item{.data}{data.frame with some list columns}

\item{id_col}{the id column to rejoin unnested list-columns with non-list columns}
}
\value{
a data frame
}
\description{
Find and unnest any list-columns in a data frame.
}

