% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_cols.R
\name{unnest_recursive}
\alias{unnest_recursive}
\title{Unnests a a data frame of list-columns}
\usage{
unnest_recursive(x, id_col = "id")
}
\arguments{
\item{x}{data.frame with one \code{N} columns, where one colum is the \code{id_col} and the
\code{N-1} are all list-columns}

\item{id_col}{id column which will be used to join unnested lists back together}
}
\value{
a data frame
}
\description{
Unnests a a data frame of list-columns
}

