% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unpack_tb}
\alias{unpack_tb}
\title{Unpack a TeachBoost response object (or a list of such objects) into a data frame}
\usage{
unpack_tb(x, unnest_cols = TRUE)
}
\arguments{
\item{x}{the TeachBoost response object (returned by \code{\link{get_tb}} to be unpacked.}

\item{unnest_cols}{should nested list-columns be unnested, default is \code{TRUE}}
}
\value{
a data frame
}
\description{
Tries to create a relatively tidy data frame, though it will maintain nested
structures as list-columns.
}
\examples{
\dontrun{
x <- get_tb("users")

x_df <- unpack_tb(x)
}
}

